% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycenter.R
\name{gaussian_mean}
\alias{gaussian_mean}
\title{Update for the GMM variance part of the MAW barycenter}
\usage{
gaussian_mean(V, w, Sigma0)
}
\description{
Update for the GMM variance part of the MAW barycenter
}
