% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAWdist.R
\name{Mawdist}
\alias{Mawdist}
\title{Compute the MAW distance between two GMMs}
\usage{
Mawdist(d, supp1, supp2, w1, w2, method = "Schur")
}
\arguments{
\item{d}{An integer which is the dimension of the original data}

\item{supp1}{A matrix that stores the mean and covariance for the first GMM.
More specifically, the dimension is (d+d\emph{d)x(J_1+J_2+...+J_N),
where d is the dimension of the original data.
For each column of supp, the first d rows store the mean vector,
#' and the last d}d rows store the vectorized covariance matrix for each component.}

\item{supp2}{A matrix that stores the mean and covariance for the first GMM.
The format is the same as the details in supp1.}

\item{w1}{A vector of weights for the first GMM that sum to 1.}

\item{w2}{A vector of weights for the second GMM that sum to 1.}

\item{method}{Methods to compute the square root of a matrix. Default is "Schur" for
Schur decomposition. Users can select "PRACMA" indicates use pracma packages directly for computation
or "Eigen" that uses eigen-decomposition.}
}
\value{
Returns a list that contains MAW distance between two GMMs as well as
the optimal transport between them.
}
\description{
The main steps of this procedure are outlined below. For a more detailed
description of the methodology, please see Lin L, Shi W, Ye J, Li J. (2023):
\doi{10.1111/biom.13630}
}
\examples{
\dontrun{
library(RMAW)
system.file("extdata", "mouse_2.mat", package = "myfirstpackage")

# mouse_2 is a an example data file containing 8 GMMs mean and variance
stride <- mouse_2$stride
supp1 <- mouse_2$supp
result <- Mawdist(d = 2,
                  supp1 = mouse$supp[,1:stride[1]],
                  supp2 = mouse$supp[,(stride[1]+1):(stride[1]+stride[2])],
                  w1 = mouse$ww[1:stride[1]],
                  w2 = mouse$ww[(stride[1]+1):(stride[1]+stride[2])])


}

}
\references{
Lin L, Shi W, Ye J, Li J. (2023)
Multisource single-cell data integration by MAW barycenter for Gaussian mixture models.
Biometrics, 79, 866–877. https://doi.org/10.1111/biom.13630
}
\concept{MAW}
