% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_2}
\alias{mouse_2}
\title{GMM fitting component data coming from mouse atlas}
\format{
A list containing the GMM info of 4 subjects onto 2 dimensions:
\describe{
\item{ww}{Concatenated weights of each component across all subjects}
\item{supp}{Mean and Variance info of each component across all subjects}
\item{stride}{Number of components for each GMM corresponding to each subject}
\item{m}{Targeted number of components for barycenter computation}
}

A list containing the GMM info of 3 subjects onto 2 dimensions:
\describe{
\item{ww}{Concatenated weights of each component between both subjects}
\item{supp}{Mean and Variance info of each component between the two subjects}
\item{stride}{Number of components for each GMM corresponding to each subject}
\item{m}{Targeted number of components for barycenter computation}
\item{d}{Number of dimensions}
}
}
\usage{
mouse_2

mouse_2
}
\description{
GMM fitting component data coming from mouse atlas

GMM fitting component data coming from COVID study
}
\keyword{datasets}
